+++ luag-lib/v1/CHANGELOG.md
@@ -5,3 +5,4 @@
     It is optional, and its default value is `0xffffff` (white).
 2. `settransparent` can now be called without an argument.
     This will remove the transparent color (aka color key).
+3. TODO (mouse functions)